% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecms.R
\name{calculate_ECMS}
\alias{calculate_ECMS}
\title{Calculate Entity Cluster Matching Score}
\usage{
calculate_ECMS(
  data,
  clustering,
  feature_importance,
  categorical_columns,
  cluster_name,
  instance
)
}
\arguments{
\item{data}{the data that was used in order to generate the clusters}

\item{feature_importance}{the importance of features as calculated by \link[clusterExplainR]{calculate_feature_importance_cluster}}

\item{categorical_columns}{all column indexes that are of categorical/binary/ordinal type}

\item{cluster_name}{the name of the cluster for which ECMS should be calculated}

\item{instance}{an instance from the dataset, use one row from table of all instances}

\item{clusterings}{the clusters that have been assigned to the data instances}
}
\value{
The Entity cluster matching score as well as individual scores per feature:
 \item{ECMS_total}{the calculated ECMS}
 \item{EMS}{a table that showns the indicidual matching scores per feature as well as feature importances}
}
\description{
Calculates the Entity Cluster Matching Scores (ECMS) which indicate how well an instance fits a specific cluster.
An instance with a high ECMS fits the cluster better than an instance with a lower ECMS.
\deqn{ECMS(x,c) = \frac{ \sum_{f \in F} p(x_f,c_f) * FIS(f,c)}{ \sum_{f \in F} FIS(f,c)}}
}
\examples{
data <- clusterExplainR::example_mocked_clustering
data_wo_clustering <- data[c(1,2)]
clustering <- data[[3]]
# indicate which columns are categorical and which are numerical
numerical_col_idx <- c(2)
categorical_col_idx <- c(1)
# calculate feature importance scores
cluster_of_interest <- 1
instance_of_interest <- data_wo_clustering[1,]
fis <- calculate_feature_importance_cluster(data,cluster_of_interest,numerical_col_idx,categorical_col_idx)
ecms <- calculate_ECMS(data_wo_clustering, clustering, fis, categorical_col_idx, cluster_of_interest, instance_of_interest)
# the overall matching score of the instance to the cluster
ecms$ECMS_total
# shows the individual matching scores per feature
ecms$EMS
}
\keyword{internal}
