% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{calculate_numeric_featureimportance_continious_entropy}
\alias{calculate_numeric_featureimportance_continious_entropy}
\title{Cluster Feature Importance for continious Features}
\usage{
calculate_numeric_featureimportance_continious_entropy(
  data_population,
  column_name,
  cluster_column_name,
  cluster_name
)
}
\arguments{
\item{data_population}{The dataset including the clustering for all clusters of interest}

\item{column_name}{the column name of the feature, of whom the feature importance should be assessed}

\item{cluster_column_name}{the name of the column that represents the clustering}

\item{cluster_name}{the name of the cluster for which the feature importance should be calculated}
}
\value{
the feature importance as a floating number between 0 and 1, where higher values indicate higher feature importance
}
\description{
Calculates the feature-importance for a continous feature by comparing the continious entropies of all instances with the instances that fit a specific cluster.
Values are kept between 0 and 1, where 1 indicates a high feature importance, while 0 indicates a no feature importance, since the entropies are equal.
\deqn{FE_{f | c} = 1- ( \frac{h_{f | c}}{h_f})}
In order to compare continious entropies over different features, continious entropy is estimated using 100 points within the feature value range
and using the NlinTS::entropy_cont function to calculate the continous entropy
}
\examples{
data <- clusterExplainR::example_mocked_clustering
# Cluster 3 contains only individuals with a low weight, while the other clusters show similar distibutions
calculate_numeric_featureimportance_continious_entropy(data, 'Weight', 'Cluster', 3)
calculate_numeric_featureimportance_continious_entropy(data, 'Weight', 'Cluster', 1)
calculate_numeric_featureimportance_continious_entropy(data, 'Weight', 'Cluster', 2)
}
\keyword{internal}
