% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_basics.R
\name{categoric_feature_rule}
\alias{categoric_feature_rule}
\title{Categorical Rule}
\usage{
categoric_feature_rule(column_name, values)
}
\arguments{
\item{column_name}{Name of the column for the Rule}

\item{values}{list of categorical values that satisfy the Rule}
}
\value{
A rule on categorical features that can be applied to a dataframe:
 \item{type}{The type of the rule}
 \item{column}{The column name the rule is applied to}
 \item{values}{The values that satisfy the rule}
 \item{apply(dataframe)}{unction to apply the defined rule to a dataframe. Returns a dataframe with objects only satisfying the rule}
 \item{verbalize}{verbalization (Human readable representation) of the rule}
}
\description{
Categorical Rule
}
\examples{
test <- data.frame(cat = c('a','b','c','a','b','c'), num = c(1,1,2,2,3,3), cluster=c(1,2,1,2,1,2))
cat_rule <- categoric_feature_rule('cat', c('a'))
print(cat_rule$verbalize)
cat_rule$apply(test)
}
\keyword{internal}
