% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecms.R
\name{categoric_matching_score}
\alias{categoric_matching_score}
\title{Matching Score for categorical/descrete variables}
\usage{
categoric_matching_score(data_clustered, column_name, cluster_name, value)
}
\arguments{
\item{data_clustered}{the clustered Dataset, expects the column with the clustering to be named 'Cluster'}

\item{column_name}{the feature/column for which the matching score should be calculates}

\item{cluster_name}{the name of the cluster for which the individual matching score should be generated}

\item{value}{a value for the desired feature}
}
\value{
a matching score of a value for a given feature/column
}
\description{
The Matching Score evaluates how well a value matches the cluster values.
Basic idea, comparing the probability of a given value for which matching score should be calculated, with highest probability of a single value.
\deqn{MS_{x|f_c} = P(x | f_c)/ max_{v \in f_c}(P(v | f_c))}
 Takes a value range [0,1]
}
\examples{
data <- clusterExplainR::example_mocked_clustering
# In Cluster 1 there are 7 females and 1 male
# for 'Female': MS = (7/8)/(7/8) = 1 (since Female is most probable value)
categoric_matching_score(data,'Gender',1,'Female')
# for 'Male': MS = (1/8)/(7/8) = 0.1428571 (since Male is a less probable value)
categoric_matching_score(data,'Gender',1,'Male')
}
\keyword{internal}
