% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_basics.R
\name{chain_rules_disjunction}
\alias{chain_rules_disjunction}
\title{Disjunctional Rule Chaining}
\usage{
chain_rules_disjunction(rules)
}
\arguments{
\item{rules}{a set of rules that should be combined in disjunction}
}
\value{
A rule composed of multiple rules:
 \item{type}{The type of the rule}
 \item{rules}{A list of the individual rules that were combined}
 \item{apply(dataframe)}{Function to apply the combined rule to a dataframe. Returns a dataframe with objects only satisfying the rule}
 \item{verbalize}{Verbalizationof the rule (Human readable representation)}
}
\description{
Combining multiple rules either numerical or categorical or combined into a single rule using disjunction
}
\examples{
test <- data.frame(cat = c('a','b','c','a','b','c'), num = c(1,1,2,2,3,3), cluster=c(1,2,1,2,1,2))
cat_rule <- categoric_feature_rule('cat', c('a'))
num_rule <- numeric_feature_rule('num',1,1)
rules <- list(cat_rule, num_rule)
chain_rules_disjunction(rules)$apply(test)
}
\keyword{internal}
