% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterExplainR.R
\docType{package}
\name{clusterExplainR}
\alias{clusterExplainR-package}
\alias{clusterExplainR}
\title{clusterExplainR}
\description{
The basic idea is the creation of explanations for mixed data clusters
purely from feature value distributions within and between clusters.
These explanations consist of:
}
\details{
1. Global Explanations: Feature Importance Scores and Feature Visualizations

 2. Local Explanations: Feature Importance Scores, Visualizations, Cluster Representatives and Rules
}
\author{
\strong{Maintainer}: Jonas Amling \email{j.amling@gmx.de}

Authors:
\itemize{
  \item Stephan Scheele \email{stephan.scheele@uni-bamberg.de}
  \item Moritz Lang \email{moritz.lang@dab-gmbh.de}
}

}
