% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{compute_discrete_entropy}
\alias{compute_discrete_entropy}
\title{Calculates the Shannon Entropy}
\usage{
compute_discrete_entropy(probabilities)
}
\arguments{
\item{probabilities}{a list of probabilities, representing the probabilities of distinct feature values}
}
\value{
The Shannon Entropy H
}
\description{
Calculates the Shannon Entropy for discrete data.
\deqn{H(X) = - \sum_{x \in X} p(x) * \log_2(p(x)) }
For continuous data the continuous entropy formula has to be used
}
\examples{
y = c(.4,.2,.1,.1)
compute_discrete_entropy(y)
}
\keyword{internal}
