% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{describe_feature}
\alias{describe_feature}
\title{Verbal description of a feature present in a cluster}
\usage{
describe_feature(
  data,
  clustering,
  categorical_columns,
  cluster_name,
  column_name
)
}
\arguments{
\item{data}{the data that was used in order to generate the clusters}

\item{clustering}{the clusters that have been assigned to the data instances}

\item{cluster_name}{the cluster name for which the local scores should be calculated}

\item{column_name}{the name of the column/feature to be described}
}
\value{
verbal description
}
\description{
Generates a verbal description for a feature. Descriptions are dependent on the type of the feature.
}
\examples{
data <- clusterExplainR::example_mocked_clustering
data_wo_clustering <- data[c(1,2)]
clustering <- data[3]
# indicate which columns are categorical and which are numerical
numerical_col_idx <- c(2)
categorical_col_idx <- c(1)
cat(describe_feature(data_wo_clustering, clustering, categorical_col_idx, 1, 'Gender'))
cat(describe_feature(data_wo_clustering, clustering,categorical_col_idx, 1, 'Weight'))
}
\keyword{internal}
