% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{describe_numerical_feature}
\alias{describe_numerical_feature}
\title{Verbal Description of a numerical feature}
\usage{
describe_numerical_feature(
  data,
  clustering,
  cluster_name,
  column_name,
  percentage = 0.9,
  num_digits = 1
)
}
\arguments{
\item{data}{the data that was used in order to generate the clusters}

\item{clustering}{the clusters that have been assigned to the data instances}

\item{cluster_name}{the cluster name for which the local scores should be calculated}

\item{column_name}{the name of the column/feature to be described}

\item{percentage}{range that should be shown additionally to max and min values, e.g. .9 -> 90 percent range}

\item{num_digits}{rounding of digits}
}
\value{
verbal description of a numerical feature
}
\description{
Verbal Description of a numerical feature
}
\examples{
data <- clusterExplainR::example_mocked_clustering
data_wo_clustering <- data[c(1,2)]
clustering <- data[3]
# generate decription of weight for cluster 1
cat(describe_numerical_feature(data_wo_clustering, clustering, 1, 'Weight'))
cat(describe_numerical_feature(data_wo_clustering, clustering, 1, 'Weight',percentage = .7 ,num_digits = 2))
}
\keyword{internal}
