% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{describe_numerical_feature_simplified}
\alias{describe_numerical_feature_simplified}
\title{Verbal Description of a numerical feature simplified}
\usage{
describe_numerical_feature_simplified(
  data,
  clustering,
  cluster_name,
  column_name
)
}
\arguments{
\item{data}{the data that was used in order to generate the clusters}

\item{clustering}{the clusters that have been assigned to the data instances}

\item{cluster_name}{the cluster name for which the local scores should be calculated}

\item{column_name}{the name of the column/feature to be described}
}
\value{
simplified verbal description of a numerical feature
}
\description{
A simple description by categorizing a feature value into 5 bins, very low, low, medium, high ,and very high
}
\examples{
data <- clusterExplainR::example_mocked_clustering
data_wo_clustering <- data[c(1,2)]
clustering <- data[3]
# simpified description of weight and cluster 1
cat(describe_numerical_feature_simplified(data_wo_clustering, clustering, 1, 'Weight'))
}
\keyword{internal}
