% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_basics.R
\name{evaluate_rule_application_in_conjunction}
\alias{evaluate_rule_application_in_conjunction}
\title{Evaluate Conjunction of Rules}
\usage{
evaluate_rule_application_in_conjunction(data, rules, target_column, target)
}
\arguments{
\item{data}{some dataframe on which the rules should be evaluated}

\item{rules}{a set of rules that should be evaluated}

\item{target_column}{name of the column that should be predicted}

\item{target}{the target value of the target column that should be predicted}
}
\value{
a dataframe that consists of:
 \item{Rule}{The verbalization of the rule, added latest}
 \item{Accuracy}{The Accuracy of the Rule, how many instances are correctly predicted}
 \item{Coverage}{Coverage of the Rule, how much the rule covers the targeted instances}
 \item{TP}{True Positive predictions}
 \item{FP}{False Positive predictions}
 \item{TN}{True Negative predictions}
 \item{FN}F{alse Negative predictions}
}
\description{
Evaluates how well a set of rules predicts a certain target. Works as follows: First Rule is evaluated
in terms of Accuracy and Coverage using TP,FP,TN and FN.
The next rule from the rule set is added in conjunction and evaluated until no more rules are left.
However, if ever an accuracy of 1.0 is reached the evaluation terminates since the target is already optimally predicted
}
\examples{
test <- data.frame(cat = c('a','b','c','a','b','c'), num = c(1,1,2,2,3,3), cluster=c(1,2,1,2,1,2))
cat_rule <- categoric_feature_rule('cat', c('a'))
num_rule <- numeric_feature_rule('num',1,1)
rules <- list(cat_rule, num_rule)
evaluate_rule_application_in_conjunction(test,rules,'cluster',1)
}
\keyword{internal}
