% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterExplainR.R
\name{explain_cluster}
\alias{explain_cluster}
\title{Local Cluster Explanations}
\usage{
explain_cluster(
  data,
  clusterings,
  numerical_columns,
  categorical_columns,
  cluster,
  generate_rules = TRUE,
  generate_prototypes = TRUE
)
}
\arguments{
\item{data}{the data that was used in order to generate the clusters}

\item{clusterings}{the clusters that have been assigned to the data instances}

\item{numerical_columns}{all columns indexes that are of numerical/continuous type}

\item{categorical_columns}{all column indexes that are of categorical/binary/ordinal type}

\item{cluster}{the cluster name for which the local scores should be calculated}

\item{generate_rules}{boolean to indicate if rules for individual clusters should be generated}

\item{generate_prototypes}{boolean to indicate if prototypes via ECMS for individual clusters should be generated}
}
\value{
An explanation of a cluster based on feature importance that compare the cluster with a population:
 \item{feature_importance}{returns a table that contains Cluster_id (name of the cluster), Column_name (name of the column -- the feature), Importance_score (the actual feature importance). Wraps: \link[clusterExplainR]{calculate_feature_importance_cluster}}
 \item{inspect_feature(column_name)}{function that can be called with a column name and returns a verbalization, the feature importance as well as a visualization (plot)}
 \item{top_n_features(number_of_features)}{function that returns a visualization of the n most important features. For each feature a subplot is created that shows the feature values the cluster takes in combination with the population values}
 \item{ECMS}{a vector that contains one entry per data row representing th EntityClusterMatchingScore (higher values suggest good fit into a cluster) \link[clusterExplainR]{calculate_ECMS}}
 \item{cluster_represenatative(number_of_entities)}{returns a list of cluster representatives, the n representatives with the highest ECMS that were assigned the cluster.}
 \item{rule}{if rules are generated, contains a descriptive rule of the cluster. For details, see \link[clusterExplainR]{generate_min_set_rules}}
}
\description{
Explains a cluster by ordering the features according to their importance.
The local feature importance scores show which features are important for
a given cluster, and are calculated by comparing the entropy for a feature
over all clusters vs within a specific cluster.
If the entropy within a cluster is lower than within the population
(all clusters combined) the feature importance is high.
For categorical features the classical discrete entropy is used,
 while for continuous features the discrete entropy is estimated.
Based on local feature importance, the n-most important features can be
visualized. Additionally, based on the importance scores entities can be
assigned an EntityClusterMatchingScore. Therefore, for each entity it is
compared how well its feature value fits within the cluster distribution and
is further weighted by the feature importance score. Ergo, a cluster
representative should be in the center of the cluster within those
features that are more important.Moreover a conjunction rule set is created
that describes the feature. Rules are created to optimize the F1 Score in
order to provide Accuracy and Coverage.
}
\examples{
data <- clusterExplainR::example_mocked_clustering
data_wo_clustering <- data[c(1,2)]
clustering <- data[[3]]
# indicate which columns are categorical and which are numerical
numerical_col_idx <- c(2)
categorical_col_idx <- c(1)
# explain cluster 3
explanation_3 <- explain_cluster(data_wo_clustering, clustering,numerical_col_idx,categorical_col_idx,3)
head(explanation_3$feature_importance,2)
explanation_3$inspect_feature('Weight')$plot
explanation_3$top_n_features(2)
explanation_3$cluster_representative(2)
cat(explanation_3$rule$verbalization)
}
