% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{generate_example_dataset}
\alias{generate_example_dataset}
\title{Example Data generation}
\usage{
generate_example_dataset()
}
\value{
an example dataset
}
\description{
Example Data generation
}
\examples{
dataset <- generate_example_dataset()

}
\keyword{internal}
