% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_basics.R
\name{generate_rule_based_on_F1}
\alias{generate_rule_based_on_F1}
\title{Rule Generation based on F1}
\usage{
generate_rule_based_on_F1(
  data,
  clustering,
  categoric_columns,
  column_name,
  cluster_name
)
}
\arguments{
\item{data}{the data that was used in order to generate the clusters}

\item{clustering}{the clusters that have been assigned to the data instances}

\item{categoric_columns}{all column indexes that are of categorical/binary/ordinal type}

\item{column_name}{the column that should be used to create the optimal rule}

\item{cluster_name}{the name of the cluster that should be predicted}
}
\value{
a rule either categorical or numerical depending of the type of the feature
}
\description{
Find the best possible rule for a given feature to predict a specific cluster, using the F1 score to evaluate possible rules.
For categorical features all permutations of possible values that can predict the cluster are used and the best one is selected.
For numerical features all possible ranges in 5% steps are evaluated.
The rules are selected based in the F1 Score in order to balance accuracy and coverage by using the harmonic mean of precision and recall (F1).
}
\examples{
data <- clusterExplainR::example_mocked_clustering
clustering <- data$Cluster
rule_cat <- generate_rule_based_on_F1(data[c(1,2)],clustering,c(1),'Gender',1)
rule_cat$verbalize
rule_num <- generate_rule_based_on_F1(data[c(1,2)],clustering,c(1),'Weight',1)
rule_num$verbalize
}
\keyword{internal}
