% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_basics.R
\name{numeric_feature_rule}
\alias{numeric_feature_rule}
\title{Numeric Rule}
\usage{
numeric_feature_rule(column_name, min_val, max_val)
}
\arguments{
\item{column_name}{Name of the column for the Rule}

\item{min_val}{the minimum that is included in the rule}

\item{max_val}{the maximum that is included in the rule}
}
\value{
A rule on numeric features that can be applied to a dataframe:
 \item{type}{The type of the rule}
 \item{column}{The column name the rule is applied to}
 \item{min}{The min value that satisfy the rule}
 \item{max}{The min value that satisfy the rule}
 \item{apply(dataframe)}{Function to apply the defined rule to a dataframe. Returns a dataframe with objects only satisfying the rule}
 \item{verbalize}{Verbalization (Human readable representation) of the rule}
}
\description{
Numeric Rule
}
\examples{
test <- data.frame(cat = c('a','b','c','a','b','c'), num = c(1,1,2,2,3,3), cluster=c(1,2,1,2,1,2))
num_rule <- numeric_feature_rule('num',1,1)
print(num_rule$verbalize)
num_rule$apply(test)
}
\keyword{internal}
