% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecms.R
\name{numeric_matching_score}
\alias{numeric_matching_score}
\title{Matching Score for numerical/continious variables}
\usage{
numeric_matching_score(data_clustered, column_name, cluster_name, value)
}
\arguments{
\item{data_clustered}{the clustered Dataset, expects the column with the clustering to be named 'Cluster'}

\item{column_name}{the feature/column for which the matching score should be calculates}

\item{cluster_name}{the name of the cluster for which the individual matching score should be generated}

\item{value}{a value for the desired feature}
}
\value{
a matching score of a value for a given feature/column
}
\description{
The Matching Score evaluates how well a value matches the cluster values.
Basic idea, comparing the probability of a given value for which matching score should be calculated, with highest probability of a single value.
For continious data the probabilities are estimated by approximating the density distributions.
\deqn{MS_{x|f_c} = d(x)/ max_{v \in f_c}(d(v))}
 Takes a value range [0,1]
}
\examples{
data <- clusterExplainR::example_mocked_clustering
numeric_matching_score(data,'Weight',2,50)
numeric_matching_score(data,'Weight',2,68)
numeric_matching_score(data,'Weight',2,90)
}
\keyword{internal}
