% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_basics.R
\name{rule_is_in_rulelist}
\alias{rule_is_in_rulelist}
\title{Check if a rule is contained in a list of rules}
\usage{
rule_is_in_rulelist(rule, rule_list)
}
\arguments{
\item{rule}{a rule as implemented}

\item{rule_list}{a list of rules}
}
\value{
Boolean
}
\description{
Checks if a rule is contained in a list of rules, by checking if the verablizations are the same.
This works since the verbalizations of a rule are automatically generated.
Does work for composite rules but order matters.
}
\examples{
cat_rule <- categoric_feature_rule('cat', c('a'))
num_rule <- numeric_feature_rule('num',1,1)
rules <- list(cat_rule, num_rule)
rule_is_in_rulelist(num_rule, rules)
rule_is_in_rulelist(categoric_feature_rule('cat', c('b')), rules)
}
\keyword{internal}
